Get rid of setup.py, setup_gae.py, install_third_party.py
