+++ scripts/backend_test_shards.json
@@ -221,7 +221,6 @@
         "core.storage.suggestion.gae_models_test",
         "core.storage.auth.gae_models_test",
         "scripts.linters.html_linter_test",
-        "scripts.setup_gae_test",
         "scripts.linters.general_purpose_linter_test",
         "core.storage.app_feedback_report.gae_models_test",
         "core.jobs.batch_jobs.model_validation_jobs_test",
@@ -283,10 +282,8 @@
         "core.storage.recommendations.gae_models_test",
         "scripts.linters.other_files_linter_test",
         "scripts.install_dependencies_json_packages_test",
-        "scripts.install_third_party_test",
         "scripts.check_frontend_test_coverage_test",
         "core.platform.models_test",
-        "scripts.setup_test",
         "core.storage.feedback.gae_models_test",
         "scripts.pre_commit_hook_test",
         "core.storage.activity.gae_models_test",
