+++ scripts/pre_push_hook_test.py
@@ -611,7 +611,7 @@ class PrePushHookTests(test_utils.GenericTestBase):
                 '\n',
                 'Please fix these discrepancies by editing the '
                 '`requirements.in`\nfile or running '
-                '`scripts.install_third_party` to regenerate\nthe '
+                '`scripts.install_third_party_libs` to regenerate\nthe '
                 '`third_party/python_libs` directory.\n\n'
             ])
 
@@ -648,7 +648,7 @@ class PrePushHookTests(test_utils.GenericTestBase):
                 '\n',
                 'Please fix these discrepancies by editing the '
                 '`requirements.in`\nfile or running '
-                '`scripts.install_third_party` to regenerate\nthe '
+                '`scripts.install_third_party_libs` to regenerate\nthe '
                 '`third_party/python_libs` directory.\n\n'
             ])
